+++ ios/Classes/Views/CupertinoSearchBarPlatformView.swift
@@ -137,6 +137,10 @@ class CupertinoSearchBarPlatformView: NSObject, FlutterPlatformView {
         }
     }
 
+    deinit {
+        channel.setMethodCallHandler(nil)
+    }
+
     func view() -> UIView {
         return container
     }